+++ rust/src/frb_generated.rs
@@ -38,7 +38,7 @@ flutter_rust_bridge::frb_generated_boilerplate!(
     default_rust_auto_opaque = RustAutoOpaqueMoi,
 );
 pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_VERSION: &str = "2.12.0";
-pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -806981178;
+pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = 907895331;
 
 // Section: executor
 
@@ -175,6 +175,39 @@ fn wire__crate__api__engine_api__engine_device_id_impl(
             deserializer.end();
             transform_result_sse::<_, ()>((move || {
                 let output_ok = Result::<_, ()>::Ok(crate::api::engine_api::engine_device_id())?;
+                Ok(output_ok)
+            })())
+        },
+    )
+}
+fn wire__crate__api__engine_api__engine_fountain_encode_impl(
+    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
+    rust_vec_len_: i32,
+    data_len_: i32,
+) -> flutter_rust_bridge::for_generated::WireSyncRust2DartSse {
+    FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec, _>(
+        flutter_rust_bridge::for_generated::TaskInfo {
+            debug_name: "engine_fountain_encode",
+            port: None,
+            mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync,
+        },
+        move || {
+            let message = unsafe {
+                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
+                    ptr_,
+                    rust_vec_len_,
+                    data_len_,
+                )
+            };
+            let mut deserializer =
+                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_data_base64 = <String>::sse_decode(&mut deserializer);
+            let api_block_size = <u32>::sse_decode(&mut deserializer);
+            deserializer.end();
+            transform_result_sse::<_, ()>((move || {
+                let output_ok = Result::<_, ()>::Ok(
+                    crate::api::engine_api::engine_fountain_encode(api_data_base64, api_block_size),
+                )?;
                 Ok(output_ok)
             })())
         },
@@ -1340,46 +1373,46 @@ fn pde_ffi_dispatcher_primary_impl(
             rust_vec_len,
             data_len,
         ),
-        14 => wire__crate__api__engine_api__engine_init_impl(port, ptr, rust_vec_len, data_len),
-        17 => wire__crate__api__engine_api__engine_pause_transfer_impl(
+        15 => wire__crate__api__engine_api__engine_init_impl(port, ptr, rust_vec_len, data_len),
+        18 => wire__crate__api__engine_api__engine_pause_transfer_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        18 => wire__crate__api__engine_api__engine_resume_transfer_impl(
+        19 => wire__crate__api__engine_api__engine_resume_transfer_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        19 => wire__crate__api__engine_api__engine_save_settings_impl(
+        20 => wire__crate__api__engine_api__engine_save_settings_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        21 => wire__crate__api__engine_api__engine_send_clipboard_impl(
+        22 => wire__crate__api__engine_api__engine_send_clipboard_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        22 => {
+        23 => {
             wire__crate__api__engine_api__engine_send_files_impl(port, ptr, rust_vec_len, data_len)
         }
-        23 => wire__crate__api__engine_api__engine_set_device_name_impl(
+        24 => wire__crate__api__engine_api__engine_set_device_name_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        26 => wire__crate__api__engine_api__engine_stop_impl(port, ptr, rust_vec_len, data_len),
-        28 => {
+        27 => wire__crate__api__engine_api__engine_stop_impl(port, ptr, rust_vec_len, data_len),
+        29 => {
             wire__crate__api__engine_api__engine_subnet_scan_impl(port, ptr, rust_vec_len, data_len)
         }
-        35 => wire__crate__api__simple__init_app_impl(port, ptr, rust_vec_len, data_len),
-        36 => wire__crate__api__types__platform_capabilities_default_impl(
+        36 => wire__crate__api__simple__init_app_impl(port, ptr, rust_vec_len, data_len),
+        37 => wire__crate__api__types__platform_capabilities_default_impl(
             port,
             ptr,
             rust_vec_len,
@@ -1398,35 +1431,36 @@ fn pde_ffi_dispatcher_sync_impl(
     // Codec=Pde (Serialization + dispatch), see doc to use other codecs
     match func_id {
         4 => wire__crate__api__engine_api__engine_device_id_impl(ptr, rust_vec_len, data_len),
-        5 => wire__crate__api__engine_api__engine_generate_pin_impl(ptr, rust_vec_len, data_len),
-        6 => wire__crate__api__engine_api__engine_generate_qr_invitation_impl(
+        5 => wire__crate__api__engine_api__engine_fountain_encode_impl(ptr, rust_vec_len, data_len),
+        6 => wire__crate__api__engine_api__engine_generate_pin_impl(ptr, rust_vec_len, data_len),
+        7 => wire__crate__api__engine_api__engine_generate_qr_invitation_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        7 => wire__crate__api__engine_api__engine_get_devices_impl(ptr, rust_vec_len, data_len),
-        8 => wire__crate__api__engine_api__engine_get_events_impl(ptr, rust_vec_len, data_len),
-        9 => wire__crate__api__engine_api__engine_get_local_ips_impl(ptr, rust_vec_len, data_len),
-        10 => wire__crate__api__engine_api__engine_get_progress_impl(ptr, rust_vec_len, data_len),
-        11 => wire__crate__api__engine_api__engine_get_stats_impl(ptr, rust_vec_len, data_len),
-        12 => wire__crate__api__engine_api__engine_get_streams_impl(ptr, rust_vec_len, data_len),
-        13 => wire__crate__api__engine_api__engine_get_transfers_impl(ptr, rust_vec_len, data_len),
-        15 => wire__crate__api__engine_api__engine_load_settings_impl(ptr, rust_vec_len, data_len),
-        16 => wire__crate__api__engine_api__engine_parse_qr_invitation_impl(
+        8 => wire__crate__api__engine_api__engine_get_devices_impl(ptr, rust_vec_len, data_len),
+        9 => wire__crate__api__engine_api__engine_get_events_impl(ptr, rust_vec_len, data_len),
+        10 => wire__crate__api__engine_api__engine_get_local_ips_impl(ptr, rust_vec_len, data_len),
+        11 => wire__crate__api__engine_api__engine_get_progress_impl(ptr, rust_vec_len, data_len),
+        12 => wire__crate__api__engine_api__engine_get_stats_impl(ptr, rust_vec_len, data_len),
+        13 => wire__crate__api__engine_api__engine_get_streams_impl(ptr, rust_vec_len, data_len),
+        14 => wire__crate__api__engine_api__engine_get_transfers_impl(ptr, rust_vec_len, data_len),
+        16 => wire__crate__api__engine_api__engine_load_settings_impl(ptr, rust_vec_len, data_len),
+        17 => wire__crate__api__engine_api__engine_parse_qr_invitation_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        20 => wire__crate__api__engine_api__engine_search_history_impl(ptr, rust_vec_len, data_len),
-        24 => wire__crate__api__engine_api__engine_start_stream_impl(ptr, rust_vec_len, data_len),
-        25 => wire__crate__api__engine_api__engine_state_impl(ptr, rust_vec_len, data_len),
-        27 => wire__crate__api__engine_api__engine_stop_stream_impl(ptr, rust_vec_len, data_len),
-        29 => wire__crate__api__engine_api__engine_verify_pin_impl(ptr, rust_vec_len, data_len),
-        30 => wire__crate__api__init__get_build_info_impl(ptr, rust_vec_len, data_len),
-        31 => wire__crate__api__init__get_protocol_version_impl(ptr, rust_vec_len, data_len),
-        32 => wire__crate__api__init__get_version_impl(ptr, rust_vec_len, data_len),
-        33 => wire__crate__api__simple__greet_impl(ptr, rust_vec_len, data_len),
-        34 => wire__crate__api__init__health_check_impl(ptr, rust_vec_len, data_len),
+        21 => wire__crate__api__engine_api__engine_search_history_impl(ptr, rust_vec_len, data_len),
+        25 => wire__crate__api__engine_api__engine_start_stream_impl(ptr, rust_vec_len, data_len),
+        26 => wire__crate__api__engine_api__engine_state_impl(ptr, rust_vec_len, data_len),
+        28 => wire__crate__api__engine_api__engine_stop_stream_impl(ptr, rust_vec_len, data_len),
+        30 => wire__crate__api__engine_api__engine_verify_pin_impl(ptr, rust_vec_len, data_len),
+        31 => wire__crate__api__init__get_build_info_impl(ptr, rust_vec_len, data_len),
+        32 => wire__crate__api__init__get_protocol_version_impl(ptr, rust_vec_len, data_len),
+        33 => wire__crate__api__init__get_version_impl(ptr, rust_vec_len, data_len),
+        34 => wire__crate__api__simple__greet_impl(ptr, rust_vec_len, data_len),
+        35 => wire__crate__api__init__health_check_impl(ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }